pyproject.toml
[tool.poetry]
name = "projects"
version = "0.1.0"
description = ""
authors = ["Vinicius Braga <Vbragadev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.2"
watchdog = "^3.0.0"
python-dotenv = "^1.0.0"
pyinstaller = "^6.1.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"
isort = "^5.12.0"
black = "^23.11.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
